Extract state listener hook
