+++ packages/rrule-generator/src/hooks/index.ts
@@ -1,3 +1,5 @@
 export { default as useRRuleConfig } from './useRRuleConfig';
 
 export { default as useRRuleState } from './useRRuleState';
+
+export { default as useStateListener } from './useStateListener';
